+++ .gitignore
@@ -22,5 +22,5 @@ npm-debug.log*
 yarn-debug.log*
 yarn-error.log*
 
-# database
+# local database
 db.sqlite3
