+++ packages/app/cypress.config.ts
@@ -4,7 +4,7 @@ import { defineConfig } from "cypress";
 export default defineConfig({
   e2e: {
     setupNodeEvents(on, config) {
-      return require("./cypress/plugins/index.js")(on, config);
+      return require("./cypress/plugins/index.ts")(on, config);
     },
     baseUrl: "http://localhost:3000",
   },
