+++ packages/app/src/containers/HomePage/index.ts
@@ -1,0 +1,1 @@
+export { default as HomePage } from "./HomePage";
