+++ packages/db/package.json
@@ -4,7 +4,7 @@
   "main": "index.ts",
   "scripts": {
     "knex": "knex",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "echo \"Error: no test specified\" && exit 0"
   },
   "dependencies": {
     "knex": "^2.5.1",
